Completely remove interchange test dir
