+++ py-polars/tests/unit/interop/test_from_dataframe.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import pandas as pd
 import pyarrow as pa
 import pytest
